fixed multiplying of all widgets by UiBuilder.get_check_widget() Found out that StudioClock was created, triggered and redrawn four times in every second. After some investigations I also found out that all other widgets had four instances too and that the method to clone widget_preview from the UI file for every video preview.
My solution is to separate the clone case into a new method
load_check_widget() in which a separate ui file can be given by
parameter from which to load the widget. So I cutted out the widgets
'audio_level_adjustment' and 'widget_preview' from voctogui.ui and
pasted it into a new file widgetpreview.ui.

I hope this change is another improvement for the voctogui performance.