reformatted all pipelines for better tracibility of my further changes
As I work more and more with gstreamer pipelines I decided to prefer a 
different formatting of the gst pipeline syntax. I find it more 
intuitive, less confusing and it's better for tracing changes.

The format is the following:

--
src
! filter
! tee
  name=tee

.tee
! sink0
  name=S0
! sink1
  name=S1
--

1) As you can see '!' signs are put in front of sinks and filters 
instead of putting them behind sources and filters.
2) Parameters are indented into the next lines - each parameter in a 
single line.
3) All sources are at top indention level.
Logging still produces some indents but they are not of any useful 
meaning.