Send sd_notify when voctocore initialization is done
This enables the use with systemd's notify type. It
enables postponing the start of dependent systemd units
that pass audio or video until voctocore is ready to
listen.

To make use of the feature, replace

[Service]
Type=simple

with

[Service]
Type=notify

when starting voctocore. See man systemd.service(5) for details.

Note: This adds an additional dependency on python3-sdnotify.