accumulate repeating log messages
especially Gst sometimes logs a lot repeating messages like:

  INFO Gst: Fast copy (alpha == 1.0)

The logging of voctocore and gui is now collection repeating messages. 
The first one will be logged as usual from the second one the messages 
will be bundled like this:

  INFO Gst: Fast copy (alpha == 1.0) [REPEATING 5 TIMES]

So overall the message was logged 5 + 1 = 6 times. The collection is per 
class/module.