use preview/videocaps to configure vaapi resolution
The configuration was using preview/size to get the resolution of the 
preview encoding but this value describes the window size of the 
previews.

With this change previews/videocaps will be used.

To avoid any further confusion Config.getVideoSize() has been renamed to 
getVideoResolution() and a new method Config.getPreviewResolution() was 
added to get the encoding resolution.