added audiomixmatrix to every avsource
this implements a new way to define audio channels in the configuration:

  [source.cam1]
  audio.english = 0+1
  audio.french = 2+3

  [source.cam2]
  audio.mono = 0

This example names stereo stream 'english' consisting of channel 0 and 1 
of cam1 and 'french' consisting of channel 2 and 3.

Audio stream 'mono' will be a one channel stream from cam1:0.

Overall the output will have 5 channels with this configuration.